Create function for operators and equals clicks
This is meant to make the code more concise an less repetative. I found
myself needing a lot of the same code for the equals sign as I do for
the operators. The catch here though is that I am doing a few things
with this function- and I am not sure if this is an appropriate time to
break the rule "only do one thing in a function" or if I am creating a
risky function.