Prevent superfluous zeros after decimal in answers
I wrote this code, but while I was fixing it, I fixed the check for what
numbers to do it to, and that may have been the issue all along. I want
recorde of this incase I need to come back to it, but after this commit
I will take this out until it proves necessary.