Make it so decimal doesn't effect tooBigMsg
There was a condition for the tooBigMsg that I think was a mistake,
it looked at the digitStr when I think I wanted to see if there was a
decimal anywhere in the string. I have decided now that I want to check
for 13 characters, regardless of whether one of them is a decimal. So
I took it out entirley.