Adjust roundToDigits() to include decimal in count
This makes it to that coming out of roundToDigits() a number will only
have 13 or less characters, not just 13 or less digits plus a decimal
point. The last change I made created a bug because the roundToDigits()
had not been adjusted and therfore anything with a 13 digits and a
decimal point was not displayed.